refactor: move functions to policies directory for higher cohesion
