+++ packages/core/src/index.ts
@@ -3,4 +3,7 @@ export * from './types/sitemap.types'
 export * from './schema/page.schema'
 export * from './schema/sitemap.schema'
 export * from './utils/slug'
+export * from './utils/errors'
+export * from './policies/normalizeSlug'
+export * from './policies/stripOrder'
 export * from './parseSitemap'
